requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
certifi==2025.4.26
click==8.1.8
colorama==0.4.6
decorator==5.2.1
executing==2.2.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==8.18.1
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pika==1.3.2
prompt_toolkit==3.0.51
pure_eval==0.2.3
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
python-dotenv==1.1.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
wcwidth==0.2.13
